+++ CHANGES.md
@@ -5,7 +5,7 @@ Changes
 
 - Make `--product-path` optional. If not given, current directory is used.
 - Prevent polluting paths with `./`.
-- Fix trying to get magik file path for non-magik classes/files.
+- Don't try to convert non-Smallworld/Magik packages.
 
 2.1.1 (2023-09-26)